server side authorization added for updated and delete

+++ backend/routes/posts.js
@@ -87,9 +87,15 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
         imagePath: imagePath
     });
     console.log(post);
-    Post.updateOne({_id: req.params.id}, post).then(result =>{
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result =>{
         // console.log(result);
-        res.status(200).json({message: "Update successful"});
+        if(result.nModified > 0){
+            return res.status(200).json({message: "Update successful"});
+        }
+        res.status(401).json({
+            message: "Not authorized"
+        });
+        
     }).catch(err => {
         console.error(err);
     });
@@ -97,11 +103,15 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
 
 router.delete("/:id",checkAuth, (req,res,next)=>{
     console.log(req.params.id);
-    Post.deleteOne({_id: req.params.id}).then((result)=>{
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then((result)=>{
         console.log(result);
-        console.log("Working");
-        res.status(200).json({
-            message: "Post deleted!"
+        if(result.n > 0){
+            return res.status(200).json({
+                message: "Post deleted!"
+            });
+        }
+        return res.status(401).json({
+            message: "Not authorized"
         });
     });
 });
